+++ Documents/ReleaseNotes.md
@@ -2,7 +2,13 @@
 
 ### [<- Go Back](../README.md)
 - `11.0.0-Preview-3.7`
-   - `Sandbox`: add `Notify Events` page. Demonstrates using custom events for notifications.
+   - `IAtlasMap`. New methods.
+	- `GetVersion`. Gets the current API version number based on build number.
+	- `SetLanguage`. Sets the default language used by the map and service modules.
+	- `SetView`. Specifies which set of geopolitically disputed borders and labels are displayed on the map.
+   - `Sandbox`: add new demo pages.
+	- `NotifyEvents`. Demonstrates using custom events for notifications.
+	- `AtlasGlobal`. Demonstrates getting the map version, and setting the language and view.
 - `11.0.0-Preview-3.6`
    - `IAtlasMap`. Added new methods:
 	- `GetServiceOptions/SetServiceOptions`. Gets or sets the map service options.